完成了比赛高光接口和球员数据查询接口，但是后续考虑将高光排序移至前端实现

+++ sports-management-event/src/main/java/com/example/event/dao/basketball/BasketballMatch.java
@@ -14,10 +14,10 @@ public class BasketballMatch {
     private String matchId;
 
     @TableField("home_team_id")
-    private Long homeTeamId;
+    private Integer homeTeamId;
 
     @TableField("away_team_id")
-    private Long awayTeamId;
+    private Integer awayTeamId;
 
     @TableField("match_time")
     private LocalDateTime matchTime;
